Move BookList component to the container folder
