+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BookList from './BookList';
+import BookList from '../containers/BookList';
 import BookForm from './BookForm';
 
 const App = () => (
