+++ src/containers/BookList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 function BookList() {
   const books = useSelector(state => state);
